pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alpha-vantage>=3.0.0",
    "anthropic>=0.72.0",
    "apscheduler>=3.11.1",
    "autogen-agentchat>=0.7.5",
    "autogen-ext>=0.7.5",
    "bcrypt>=5.0.0",
    "bleak>=1.1.1",
    "cirq>=1.6.1",
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "hrv-analysis>=1.0.5",
    "ijson>=3.4.0.post0",
    "jinja2>=3.1.6",
    "kalshi-python>=2.1.4",
    "langgraph>=1.0.2",
    "markdown>=3.10",
    "matplotlib>=3.10.7",
    "mne>=1.10.2",
    "muselsl>=2.3.1",
    "networkx>=3.5",
    "numpy>=2.3.4",
    "openai>=2.7.1",
    "oura-ring>=0.3.0",
    "pandas>=2.3.3",
    "plotly>=6.4.0",
    "polar-python>=0.0.4",
    "psycopg2-binary>=2.9.11",
    "pyautogen>=0.10.0",
    "pygithub>=2.8.1",
    "pylsl>=1.17.6",
    "python-dateutil>=2.9.0.post0",
    "python-osc>=1.9.3",
    "pytrends>=4.9.2",
    "qiskit>=2.2.3",
    "qiskit-aer>=0.17.2",
    "replit-object-storage>=1.0.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "streamlit>=1.51.0",
    "stripe>=14.1.0",
    "tenacity>=9.1.2",
    "trafilatura>=2.0.0",
    "weasyprint>=66.0",
    "yfinance>=0.2.66",
]
